move into its own folder to match refactor pr and add each configuration also as a method on the instrument (as well as as a possible value of the configuration parameter)
